Working putchar/getchar
Updating the code to have a working putchar/getchar with only the low level registers.
Main change is to comment out the SDK UART init, as that entails the enablement of the Interrupt Handler of the RX FIFO, which causes there to appear as if there is never data when getchar() is called.